Cleanup + use socket
