+++ ext/node/lib.rs
@@ -5,7 +5,6 @@
 
 use std::collections::HashSet;
 use std::path::Path;
-use std::rc::Rc;
 
 use deno_core::error::AnyError;
 use deno_core::located_script_name;
@@ -14,13 +13,8 @@ use deno_core::url::Url;
 #[allow(unused_imports)]
 use deno_core::v8;
 use deno_core::v8::ExternalReference;
-use deno_core::AsyncRefCell;
-use deno_core::AsyncResult;
-use deno_core::CancelHandle;
-use deno_core::CancelTryFuture;
 use deno_core::JsRuntime;
 use deno_core::OpState;
-use deno_core::RcRef;
 use deno_fs::sync::MaybeSend;
 use deno_fs::sync::MaybeSync;
 use node_resolver::NpmResolverRc;
@@ -45,8 +39,6 @@ pub use ops::vm::VM_CONTEXT_INDEX;
 pub use polyfill::is_builtin_node_module;
 pub use polyfill::SUPPORTED_BUILTIN_NODE_MODULES;
 pub use polyfill::SUPPORTED_BUILTIN_NODE_MODULES_WITH_PREFIX;
-use tokio::io::AsyncReadExt;
-use tokio::io::AsyncWriteExt;
 
 use crate::global::global_object_middleware;
 use crate::global::global_template_middleware;
@@ -875,304 +867,4 @@ pub fn create_host_defined_options<'s>(
   let value = v8::Boolean::new(scope, true);
   host_defined_options.set(scope, 0, value.into());
   host_defined_options.into()
-}
-
-#[cfg(unix)]
-pub type RawPipeHandle = std::os::fd::RawFd;
-
-#[cfg(windows)]
-pub type RawPipeHandle = std::os::windows::raw::Handle;
-
-#[pin_project::pin_project]
-pub struct BiPipe {
-  #[pin]
-  read_end: BiPipeRead,
-  #[pin]
-  write_end: BiPipeWrite,
-}
-
-impl BiPipe {
-  pub fn from_raw(raw: RawPipeHandle) -> Result<Self, std::io::Error> {
-    let (read_end, write_end) = pipe(raw)?;
-    Ok(Self {
-      read_end,
-      write_end,
-    })
-  }
-
-  pub fn split(self) -> (BiPipeRead, BiPipeWrite) {
-    (self.read_end, self.write_end)
-  }
-
-  pub fn unsplit(read_end: BiPipeRead, write_end: BiPipeWrite) -> Self {
-    Self {
-      read_end,
-      write_end,
-    }
-  }
-}
-
-impl deno_core::Resource for BiPipeResource {
-  fn close(self: Rc<Self>) {
-    self.cancel.cancel();
-  }
-
-  fn backing_handle(self: Rc<Self>) -> Option<deno_core::ResourceHandle> {
-    Some(deno_core::ResourceHandle::from_fd_like(&self.raw_handle))
-  }
-
-  deno_core::impl_readable_byob!();
-  deno_core::impl_writable!();
-
-  // fn read(
-  //   self: Rc<Self>,
-  //   limit: usize,
-  // ) -> deno_core::AsyncResult<deno_core::BufView> {
-  //   async move {
-  //     let mut read_half = RcRef::map(self, |r| &r.read_half).borrow_mut().await;
-  //     let mut buf = vec![0u8; limit];
-  //     let read = read_half.read(&mut buf).await?;
-  //     buf.truncate(read);
-  //     Ok(buf.into())
-  //   }
-  //   .boxed_local()
-  // }
-
-  // fn write(
-  //   self: Rc<Self>,
-  //   mut buf: deno_core::BufView,
-  // ) -> deno_core::AsyncResult<deno_core::WriteOutcome> {
-  //   async move {
-  //     let mut write_half =
-  //       RcRef::map(self, |r| &r.write_half).borrow_mut().await;
-  //     let n = buf.len();
-  //     let written = write_half.write_buf(&mut buf).await?;
-  //     if written < n {
-  //       Ok(deno_core::WriteOutcome::Partial {
-  //         nwritten: written,
-  //         view: buf,
-  //       })
-  //     } else {
-  //       Ok(deno_core::WriteOutcome::Full { nwritten: written })
-  //     }
-  //   }
-  //   .boxed_local()
-  // }
-}
-
-pub struct BiPipeResource {
-  read_half: AsyncRefCell<BiPipeRead>,
-  write_half: AsyncRefCell<BiPipeWrite>,
-  cancel: CancelHandle,
-  raw_handle: RawPipeHandle,
-}
-
-impl BiPipeResource {
-  pub fn from_raw_handle(raw: RawPipeHandle) -> Result<Self, std::io::Error> {
-    let pipe = BiPipe::from_raw(raw)?;
-    let (read, write) = pipe.split();
-    Ok(Self {
-      raw_handle: raw,
-      read_half: AsyncRefCell::new(read),
-      write_half: AsyncRefCell::new(write),
-      cancel: Default::default(),
-    })
-  }
-
-  pub async fn read(
-    self: Rc<Self>,
-    data: &mut [u8],
-  ) -> Result<usize, AnyError> {
-    let mut rd = RcRef::map(&self, |r| &r.read_half).borrow_mut().await;
-    let cancel_handle = RcRef::map(&self, |r| &r.cancel);
-    Ok(rd.read(data).try_or_cancel(cancel_handle).await?)
-  }
-
-  pub async fn write(self: Rc<Self>, data: &[u8]) -> Result<usize, AnyError> {
-    let mut wr = RcRef::map(self, |r| &r.write_half).borrow_mut().await;
-    let nwritten = wr.write(data).await?;
-    wr.flush().await?;
-    Ok(nwritten)
-  }
-}
-
-#[pin_project::pin_project]
-pub struct BiPipeRead {
-  #[cfg(unix)]
-  #[pin]
-  inner: tokio::net::unix::OwnedReadHalf,
-  #[cfg(windows)]
-  #[pin]
-  inner: tokio::io::ReadHalf<tokio::net::windows::named_pipe::NamedPipeClient>,
-}
-
-impl tokio::io::AsyncRead for BiPipeRead {
-  fn poll_read(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    buf: &mut tokio::io::ReadBuf<'_>,
-  ) -> std::task::Poll<std::io::Result<()>> {
-    self.project().inner.poll_read(cx, buf)
-  }
-}
-
-#[cfg(unix)]
-impl From<tokio::net::unix::OwnedReadHalf> for BiPipeRead {
-  fn from(value: tokio::net::unix::OwnedReadHalf) -> Self {
-    Self { inner: value }
-  }
-}
-#[cfg(windows)]
-impl From<tokio::io::ReadHalf<tokio::net::windows::named_pipe::NamedPipeClient>>
-  for BiPipeRead
-{
-  fn from(
-    value: tokio::io::ReadHalf<
-      tokio::net::windows::named_pipe::NamedPipeClient,
-    >,
-  ) -> Self {
-    Self { inner: value }
-  }
-}
-
-impl tokio::io::AsyncRead for BiPipe {
-  fn poll_read(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    buf: &mut tokio::io::ReadBuf<'_>,
-  ) -> std::task::Poll<std::io::Result<()>> {
-    self.project().read_end.poll_read(cx, buf)
-  }
-}
-
-impl tokio::io::AsyncWrite for BiPipe {
-  fn poll_write_vectored(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    bufs: &[std::io::IoSlice<'_>],
-  ) -> std::task::Poll<Result<usize, std::io::Error>> {
-    self.project().write_end.poll_write_vectored(cx, bufs)
-  }
-
-  fn is_write_vectored(&self) -> bool {
-    self.write_end.is_write_vectored()
-  }
-
-  fn poll_write(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    buf: &[u8],
-  ) -> std::task::Poll<Result<usize, std::io::Error>> {
-    self.project().write_end.poll_write(cx, buf)
-  }
-
-  fn poll_flush(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-  ) -> std::task::Poll<Result<(), std::io::Error>> {
-    self.project().write_end.poll_flush(cx)
-  }
-
-  fn poll_shutdown(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-  ) -> std::task::Poll<Result<(), std::io::Error>> {
-    self.project().write_end.poll_shutdown(cx)
-  }
-}
-
-#[pin_project::pin_project]
-pub struct BiPipeWrite {
-  #[cfg(unix)]
-  #[pin]
-  inner: tokio::net::unix::OwnedWriteHalf,
-  #[cfg(windows)]
-  #[pin]
-  inner: tokio::io::WriteHalf<tokio::net::windows::named_pipe::NamedPipeClient>,
-}
-
-#[cfg(unix)]
-impl From<tokio::net::unix::OwnedWriteHalf> for BiPipeWrite {
-  fn from(value: tokio::net::unix::OwnedWriteHalf) -> Self {
-    Self { inner: value }
-  }
-}
-#[cfg(windows)]
-impl
-  From<tokio::io::WriteHalf<tokio::net::windows::named_pipe::NamedPipeClient>>
-  for BiPipeWrite
-{
-  fn from(
-    value: tokio::io::WriteHalf<
-      tokio::net::windows::named_pipe::NamedPipeClient,
-    >,
-  ) -> Self {
-    Self { inner: value }
-  }
-}
-
-impl tokio::io::AsyncWrite for BiPipeWrite {
-  fn poll_write_vectored(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    bufs: &[std::io::IoSlice<'_>],
-  ) -> std::task::Poll<Result<usize, std::io::Error>> {
-    self.project().inner.poll_write_vectored(cx, bufs)
-  }
-
-  fn is_write_vectored(&self) -> bool {
-    self.inner.is_write_vectored()
-  }
-
-  fn poll_write(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-    buf: &[u8],
-  ) -> std::task::Poll<Result<usize, std::io::Error>> {
-    self.project().inner.poll_write(cx, buf)
-  }
-
-  fn poll_flush(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-  ) -> std::task::Poll<Result<(), std::io::Error>> {
-    self.project().inner.poll_flush(cx)
-  }
-
-  fn poll_shutdown(
-    self: std::pin::Pin<&mut Self>,
-    cx: &mut std::task::Context<'_>,
-  ) -> std::task::Poll<Result<(), std::io::Error>> {
-    self.project().inner.poll_shutdown(cx)
-  }
-}
-
-#[cfg(unix)]
-pub fn pipe(
-  stream: RawPipeHandle,
-) -> Result<(BiPipeRead, BiPipeWrite), std::io::Error> {
-  // Safety: The fd is part of a pair of connected sockets created by child process
-  // implementation.
-
-  use std::os::fd::FromRawFd;
-  let unix_stream = tokio::net::UnixStream::from_std(unsafe {
-    std::os::unix::net::UnixStream::from_raw_fd(stream)
-  })?;
-  let (read, write) = unix_stream.into_split();
-  Ok((BiPipeRead { inner: read }, BiPipeWrite { inner: write }))
-}
-
-#[cfg(windows)]
-pub fn pipe(
-  handle: RawPipeHandle,
-) -> Result<(BiPipeRead, BiPipeWrite), io::Error> {
-  // Safety: We cannot use `get_osfhandle` because Deno statically links to msvcrt. It is not guaranteed that the
-  // fd handle map will be the same.
-  let pipe = unsafe {
-    tokio::net::windows::named_pipe::NamedPipeClient::from_raw_handle(
-      handle as _,
-    )?
-  };
-  let (read, write) = tokio::io::split(pipe);
-  Ok((BiPipeRead { inner: read }, BiPipeWrite { inner: write }))
 }
